requirements.txt
# Core Deep Learning
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# AI2-THOR Simulator
ai2thor==2.1.0

# AllenNLP for EmBERT
allennlp==2.10.1
allennlp-models==2.10.1

# RL Libraries
stable-baselines3==2.2.1
gymnasium==0.29.1

# Transformers and Quantization (for Qwen2-VL)
transformers>=4.37.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
qwen-vl-utils
sentencepiece==0.1.99
protobuf==3.20.3

# Computer Vision (Python 3.8 compatible)
opencv-python==4.8.0.74
pillow==9.5.0

# Scientific Computing
numpy==1.24.3
scipy==1.10.1
scikit-learn==1.3.0
pandas==2.0.3

# Visualization
matplotlib==3.7.1
seaborn==0.12.2

# Experiment Tracking
tensorboard==2.13.0
wandb==0.15.5

# Utilities
tqdm==4.66.1
pyyaml==6.0
jsonlines==3.1.0
h5py==3.9.0

# Development
jupyter==1.0.0
ipykernel==6.27.1
pytest==7.4.3
